requirements.txt
altgraph==0.17.4
certifi==2024.12.14
charset-normalizer==2.0.12
idna==3.10
importlib_metadata==8.5.0
macholib==1.16.3
packaging==24.2
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
pyTelegramBotAPI==4.14.1
python-dotenv==0.20.0
requests==2.26.0
urllib3==1.26.20
WMI==1.4.9
zipp==3.21.0